Test parsing for single and double quotes on new Linux Machine with Cursor sh
